// src/constants/enums.ts
export enum CategoryType {
  INCOME = "Income",
  EXPENSE = "Expense",
}

export enum TransactionType {
  ONE_TIME = "one-time",
  RECURRING = "recurring",
}

export enum Frequency {
  DAILY = "daily",
  WEEKLY = "weekly",
  MONTHLY = "monthly",
  QUARTERLY = "quarterly",
  YEARLY = "yearly",
}

export enum Currency {
  BIRR = "BIRR",
  USD = "USD",
  EUR = "EUR",
  GBP = "GBP",
  CAD = "CAD",
  AUD = "AUD",
  CHF = "CHF",
  CNY = "CNY",
  JPY = "JPY",
  NZD = "NZD",
  SGD = "SGD",
  HKD = "HKD",
  INR = "INR",
  MXN = "MXN",
  RUB = "RUB",
  SAR = "SAR",
}

export enum Language {
  EN = "en",
  AM = "am",
  OM = "om",
}

export enum Role {
  USER = "user",
  ADMIN = "admin",
}